Fix corruption when deleting content from a note
It turns out that the SAF may not truncate data when a Uri is opened
with a "w" mode even though that mode is supposed to "[erase] whatever
data is currently in the file". In practice, this meant that when text
was deleted from a file, either garbage data would be added to the end
of the file, or the deleted content would be included in the saved file.